etcd-runner: expose runner logic as a library
