+++ tools/functional-tester/etcd-runner/command/global.go
@@ -15,12 +15,7 @@
 package command
 
 import (
-	"fmt"
-	"log"
-	"sync"
 	"time"
-
-	"github.com/coreos/etcd/clientv3"
 
 	"github.com/spf13/cobra"
 )
@@ -28,8 +23,8 @@ import (
 var (
 	rounds                 int           // total number of rounds the operation needs to be performed
 	totalClientConnections int           // total number of client connections to be made with server
-	noOfPrefixes           int           // total number of prefixes which will be watched upon
-	watchPerPrefix         int           // number of watchers per prefix
+	numPrefixes            int           // total number of prefixes which will be watched upon
+	watchesPerPrefix       int           // number of watchers per prefix
 	reqRate                int           // put request per second
 	totalKeys              int           // total number of keys for operation
 	runningTime            time.Duration // time for which operation should be performed
@@ -40,77 +35,6 @@ var (
 type GlobalFlags struct {
 	Endpoints   []string
 	DialTimeout time.Duration
-}
-
-type roundClient struct {
-	c        *clientv3.Client
-	progress int
-	acquire  func() error
-	validate func() error
-	release  func() error
-}
-
-func newClient(eps []string, timeout time.Duration) *clientv3.Client {
-	c, err := clientv3.New(clientv3.Config{
-		Endpoints:   eps,
-		DialTimeout: time.Duration(timeout) * time.Second,
-	})
-	if err != nil {
-		log.Fatal(err)
-	}
-	return c
-}
-
-func doRounds(rcs []roundClient, rounds int) {
-	var mu sync.Mutex
-	var wg sync.WaitGroup
-
-	wg.Add(len(rcs))
-	finished := make(chan struct{}, 0)
-	for i := range rcs {
-		go func(rc *roundClient) {
-			defer wg.Done()
-			for rc.progress < rounds {
-				for rc.acquire() != nil { /* spin */
-				}
-
-				mu.Lock()
-				if err := rc.validate(); err != nil {
-					log.Fatal(err)
-				}
-				mu.Unlock()
-
-				time.Sleep(10 * time.Millisecond)
-				rc.progress++
-				finished <- struct{}{}
-
-				mu.Lock()
-				for rc.release() != nil {
-					mu.Unlock()
-					mu.Lock()
-				}
-				mu.Unlock()
-			}
-		}(&rcs[i])
-	}
-
-	start := time.Now()
-	for i := 1; i < len(rcs)*rounds+1; i++ {
-		select {
-		case <-finished:
-			if i%100 == 0 {
-				fmt.Printf("finished %d, took %v\n", i, time.Since(start))
-				start = time.Now()
-			}
-		case <-time.After(time.Minute):
-			log.Panic("no progress after 1 minute!")
-		}
-	}
-	wg.Wait()
-
-	for _, rc := range rcs {
-		rc.c.Close()
-	}
 }
 
 func endpointsFromFlag(cmd *cobra.Command) []string {
